// lib/src/lib/program/types/Authority.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from "@metaplex-foundation/beet";
/**
 * @category enums
 * @category generated
 */
export enum Authority {
  Neither,
  Creator,
  Recipient,
  Both,
}

/**
 * @category userTypes
 * @category generated
 */
export const authorityBeet = beet.fixedScalarEnum(
  Authority
) as beet.FixedSizeBeet<Authority, Authority>;